requirements.txt
fastapi
uvicorn
groq
python-dotenv
sentence-transformers
faiss-cpu
numpy<2.0.0
pydantic
